requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langgraph>=0.0.15
openai>=1.0.0

# Data processing - pin numpy to <2
numpy<2.0.0
pandas>=2.0.0
plotly>=5.18.0

# Vector store for memory
chromadb>=0.4.22
onnxruntime  # Required for chromadb embeddings

# Web search (optional)
tavily-python>=0.3.0

# Utilities
watchdog>=3.0.0